+++ app/helpers/admin/clinical_trials_helper.rb
@@ -1,0 +1,2 @@
+module Admin::ClinicalTrialsHelper
+end